requirements.txt
dateparser>=1.2
fastmcp>=2.14.4
google-genai>=0.1.0
#groq>=1.0.0
mcp>=1.0
mysql-connector-python>=8.3
nest-asyncio>=1.6.0
pydantic>=2.5
pytest>=9.0.2
python-dotenv>=1.0
requests>=2.31
setuptools>=80.10.2
sqlalchemy>=2.0
streamlit>=1.30
tenacity>=9.0